Use Unicode escape for popup closer character, fixes #5
As the CSS previously contained a literal cross character it was
essential that the CSS was identified as UTF-8 which is commonly either
done via HTTP headers or charset tag in the HTML file. @amr66 suggested
setting the `@charset "utf-8"` within the CSS file which is valid but I
felt it was simpler to use the escaped Unicode point.